# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Hoteles",
        "description": "desciption",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://gestion-hotel.test/"
        }
    ],
    "paths": {
...
        "/api/hotel": {
            "get": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Obtener lista de hoteles paginada",
                "operationId": "42e38e1ec9a77688ed78f16868c19a31",
                "responses": {
                    "200": {
                        "description": "Lista de hoteles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Crear un nuevo hotel",
                "description": "Crea un nuevo hotel con los datos proporcionados",
                "operationId": "e4cd0a9fd83fe69e1de0b7278c89a329",
                "requestBody": {
                    "description": "Datos necesarios para crear un hotel",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreHotelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Hotel creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "property": null
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/hotel/{hotel}": {
            "get": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Obtener hotel",
                "description": "Devuelve los detalles de un hotel específico.",
                "operationId": "b03182417b6160fa22c808080b9057ae",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del hotel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Editar hotel",
                "description": "Actualiza los datos de un hotel específico.",
                "operationId": "19009aa79ddf6626ff887cbd39f256d4",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutHotelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hotel actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Hotel"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Eliminar un hotel",
                "description": "Elimina un hotel específico por ID.",
                "operationId": "80235fddf92cc0a664fbc830a5880b5d",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hotel eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "ok"
                            }
                        }
                    }
                }
            }
        },
        "/api/hotel/{hotel}/habitaciones": {
            "get": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Obtener habitaciones de un hotel",
                "description": "Obtiene todas las habitaciones de un hotel.",
                "operationId": "a4ba4501067025cbc98e393e2eb90c22",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de habitaciones del hotel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Habitacion"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/hotel/{hotel}/servicio/{servicio}": {
            "post": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Asociar un servicio a un hotel",
                "operationId": "0e45435e610f6e1bde29ccff91491502",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "description": "ID del hotel al que se asociará el servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "servicio",
                        "in": "path",
                        "description": "ID del servicio a asociar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio asociado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Servicio asociado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "El servicio ya está asociado a este hotel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El servicio ya está asociado a este hotel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Desasociar un servicio de un hotel",
                "operationId": "2c46ff1898fe096db1c8efb8314754a0",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "description": "ID del hotel del que se desasociará el servicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "servicio",
                        "in": "path",
                        "description": "ID del servicio a desasociar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servicio desasociado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Servicio desasociado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "El servicio no está asociado a este hotel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El servicio no está asociado a este hotel"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/hotel/{hotel}/servicios": {
            "get": {
                "tags": [
                    "Hotel"
                ],
                "summary": "Obtener servicios de un hotel",
                "description": "Obtiene todos los servicios asociados a un hotel.",
                "operationId": "bf0df4b59e96f76a4c7a825f36400ccf",
                "parameters": [
                    {
                        "name": "hotel",
                        "in": "path",
                        "description": "ID del hotel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de servicios del hotel",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Servicio"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Hotel": {
                "title": "Hotel",
                "description": "Modelo de un hotel",
                "required": [
                    "nombre",
                    "direccion"
                ],
                "properties": {
                    "id": {
                        "description": "ID del hotel",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del hotel",
                        "type": "string"
                    },
                    "direccion": {
                        "description": "Dirección del hotel",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "Teléfono de contacto del hotel",
                        "type": "string"
                    },
                    "email": {
                        "description": "Correo electrónico del hotel",
                        "type": "string",
                        "format": "email"
                    },
                    "sitioWeb": {
                        "description": "Sitio web del hotel",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StoreHotelRequest": {
                "title": "Store Hotel Request",
                "description": "Datos requeridos para crear un nuevo hotel",
                "required": [
                    "nombre",
                    "direccion",
                    "telefono",
                    "email",
                    "sitioWeb"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del hotel (mínimo 5 caracteres, máximo 50)",
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 5
                    },
                    "direccion": {
                        "description": "Dirección del hotel (mínimo 5 caracteres, máximo 100)",
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 5
                    },
                    "telefono": {
                        "description": "Teléfono del hotel (máximo 20 caracteres)",
                        "type": "string",
                        "maxLength": 20
                    },
                    "email": {
                        "description": "Correo electrónico del hotel (mínimo 5 caracteres)",
                        "type": "string",
                        "format": "email",
                        "minLength": 5
                    },
                    "sitioWeb": {
                        "description": "Sitio web del hotel (mínimo 5 caracteres)",
                        "type": "string",
                        "minLength": 5
                    }
                },
                "type": "object"
            },
            "PutHotelRequest": {
                "title": "Put Hotel Request",
                "description": "Datos requeridos para crear un nuevo hotel",
                "required": [
                    "nombre",
                    "direccion",
                    "telefono",
                    "email",
                    "sitioWeb"
                ],
                "properties": {
                    "nombre": {
                        "description": "Nombre del hotel (mínimo 5 caracteres, máximo 50)",
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 5
                    },
                    "direccion": {
                        "description": "Dirección del hotel (mínimo 5 caracteres, máximo 100)",
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 5
                    },
                    "telefono": {
                        "description": "Teléfono del hotel (máximo 20 caracteres)",
                        "type": "string",
                        "maxLength": 20
                    },
                    "email": {
                        "description": "Correo electrónico del hotel (mínimo 5 caracteres)",
                        "type": "string",
                        "format": "email",
                        "minLength": 5
                    },
                    "sitioWeb": {
                        "description": "Sitio web del hotel (mínimo 5 caracteres)",
                        "type": "string",
                        "minLength": 5
                    }
                },
                "type": "object"
            },
            "Habitacion": {
                "title": "Habitación",
                "description": "Modelo de una habitación",
                "required": [
                    "numero",
                    "tipo",
                    "precioNoche",
                    "hotel_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID de la habitación",
                        "type": "integer"
                    },
                    "numero": {
                        "description": "Número de la habitación",
                        "type": "string"
                    },
                    "tipo": {
                        "description": "Tipo de habitación (ej. individual, doble)",
                        "type": "string"
                    },
                    "precioNoche": {
                        "description": "Precio por noche de la habitación",
                        "type": "number",
                        "format": "float"
                    },
                    "hotel_id": {
                        "description": "ID del hotel al que pertenece la habitación",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Servicio": {
                "title": "Servicio",
                "description": "Modelo de un servicio",
                "required": [
                    "nombre",
                    "descripcion",
                    "categoria"
                ],
                "properties": {
                    "id": {
                        "description": "ID del servicio",
                        "type": "integer"
                    },
                    "nombre": {
                        "description": "Nombre del servicio",
                        "type": "string"
                    },
                    "descripcion": {
                        "description": "Descripción del servicio",
                        "type": "string"
                    },
                    "categoria": {
                        "description": "Categoría del servicio",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de última actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}